# Update tflops calculation